Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
python-multipart==0.0.22
python-dotenv==1.0.1
alembic==1.18.4
cloudinary==1.44.0